Dev

See merge request polytechnique-montr-al/log2990/20251/equipe-309/LOG2990-309!142

+++ server/app/gateways/lobby/lobby.gateway.ts
@@ -1,30 +1,31 @@
-import { Player } from '@app/interfaces/Player';
+import { Player } from '@app/interfaces/player';
 import { Game } from '@app/model/database/game';
+import { AccessCodesService } from '@app/services/access-codes/access-codes.service';
+import { GameCombatService } from '@app/services/combat-manager/combat-manager.service';
+import { GameSessionService } from '@app/services/game-session/game-session.service';
 import { LobbyService } from '@app/services/lobby/lobby.service';
-import { Logger } from '@nestjs/common';
 import {
     ConnectedSocket,
     MessageBody,
     OnGatewayConnection,
     OnGatewayDisconnect,
-    OnGatewayInit,
     SubscribeMessage,
     WebSocketGateway,
     WebSocketServer,
 } from '@nestjs/websockets';
 import { Server, Socket } from 'socket.io';
 import { LobbyEvents } from './lobby.gateway.events';
-
 @WebSocketGateway({ cors: true })
-export class LobbyGateway implements OnGatewayConnection, OnGatewayDisconnect, OnGatewayInit {
+export class LobbyGateway implements OnGatewayConnection, OnGatewayDisconnect {
     @WebSocketServer()
     server: Server;
 
     constructor(
         private readonly lobbyService: LobbyService,
-        private readonly logger: Logger,
+        private readonly accessCodesService: AccessCodesService,
+        private readonly gameSessionService: GameSessionService,
+        private readonly gameCombatService: GameCombatService,
     ) {}
-
     @SubscribeMessage('requestUnavailableOptions')
     handleRequestUnavailableOptions(@MessageBody() accessCode: string, @ConnectedSocket() client: Socket) {
         const lobby = this.lobbyService.getLobby(accessCode);
@@ -43,7 +44,7 @@ export class LobbyGateway implements OnGatewayConnection, OnGatewayDisconnect, O
         const { game } = data;
         const accessCode = this.lobbyService.createLobby(game);
         client.join(accessCode);
-        this.logger.log(`Lobby created with game: ${game.name} and accessCode: ${accessCode}`);
+        this.lobbyService.addPlayerToRoom(client.id, accessCode);
         client.emit('lobbyCreated', { accessCode });
     }
 
@@ -62,8 +63,8 @@ export class LobbyGateway implements OnGatewayConnection, OnGatewayDisconnect, O
         }
         const success = this.lobbyService.joinLobby(accessCode, player).success;
         if (success) {
+            this.lobbyService.addPlayerToRoom(client.id, accessCode);
             client.join(accessCode);
-            this.logger.log(`Player ${player.name} joined lobby ${accessCode}`);
             this.lobbyService.setPlayerSocket(player.name, client.id);
             this.server.to(accessCode).emit('updatePlayers', this.lobbyService.getLobbyPlayers(accessCode));
             client.emit('joinedLobby');
@@ -83,52 +84,24 @@ export class LobbyGateway implements OnGatewayConnection, OnGatewayDisconnect, O
             client.emit('error', 'Lobby not found');
             return;
         }
-
+        this.lobbyService.addPlayerToRoom(client.id, accessCode);
         client.join(accessCode);
+        lobby.waitingPlayers = lobby.waitingPlayers.filter((wp) => this.server.sockets.sockets.has(wp.socketId));
         const unavailableAvatars = [...lobby.players.map((p) => p.avatar), ...lobby.waitingPlayers.map((wp) => wp.avatar)];
         this.server.to(client.id).emit('updateUnavailableOptions', { avatars: unavailableAvatars });
     }
 
-    @SubscribeMessage(LobbyEvents.LeaveLobby)
-    handleLeaveLobby(@MessageBody() data: { accessCode: string; playerName: string }, @ConnectedSocket() client: Socket) {
-        const { accessCode, playerName } = data;
-
-        const lobby = this.lobbyService.getLobby(accessCode);
-        if (!lobby) return;
-        const isAdminLeaving = this.lobbyService.isAdminLeaving(accessCode, playerName);
-        if (isAdminLeaving) {
-            this.server.to(accessCode).emit('adminLeft', { playerSocketId: client.id, message: "L'admin a quitté la partie, le lobby est fermé." });
-        }
-        const isLobbyDeleted = this.lobbyService.leaveLobby(accessCode, playerName);
-        lobby.waitingPlayers = lobby.waitingPlayers.filter((wp) => wp.socketId !== client.id);
-
-        if (isLobbyDeleted) {
-            this.server.to(accessCode).emit('lobbyDeleted');
-            this.server.to(accessCode).emit('updateUnavailableOptions', { avatars: [] });
-        } else {
-            const unavailableAvatars = [...lobby.players.map((p) => p.avatar), ...lobby.waitingPlayers.map((wp) => wp.avatar)];
-            this.server.to(accessCode).emit('updateUnavailableOptions', { avatars: unavailableAvatars });
-            this.server.to(accessCode).emit('updatePlayers', this.lobbyService.getLobbyPlayers(accessCode));
-            this.server.to(client.id).emit('updateUnavailableOptions', { avatars: [] });
-        }
-
-        client.leave(accessCode);
-    }
-
     @SubscribeMessage('kickPlayer')
     handleKickPlayer(@MessageBody() data: { accessCode: string; playerName: string }) {
-        this.logger.log(`Admin requested to kick player ${data.playerName} from lobby ${data.accessCode}`);
         const kickedPlayerSocketId = this.lobbyService.getPlayerSocket(data.playerName);
         if (kickedPlayerSocketId) {
             const kickedSocket = this.server.sockets.sockets.get(kickedPlayerSocketId);
             if (kickedSocket) {
-                this.handleLeaveLobby(data, kickedSocket);
+                this.handlePlayerDisconnect(kickedSocket);
             }
             this.server.to(kickedPlayerSocketId).emit('kicked', { accessCode: data.accessCode, playerName: data.playerName });
             this.lobbyService.removePlayerSocket(data.playerName);
         }
-
-        this.logger.log(`Player ${data.playerName} was kicked from lobby ${data.accessCode}`);
     }
 
     @SubscribeMessage(LobbyEvents.GetLobbyPlayers)
@@ -143,13 +116,10 @@ export class LobbyGateway implements OnGatewayConnection, OnGatewayDisconnect, O
 
     @SubscribeMessage(LobbyEvents.GetLobby)
     handleGetLobby(@MessageBody() accessCode: string, @ConnectedSocket() client: Socket) {
-        this.logger.log(`Received getLobby request for accessCode: ${accessCode}`);
         const lobby = this.lobbyService.getLobby(accessCode);
         if (lobby) {
-            this.logger.log(`Lobby found with code: ${JSON.stringify(lobby.accessCode)}`);
             client.emit('updateLobby', lobby);
         } else {
-            this.logger.log(`Lobby not found for accessCode: ${accessCode}`);
             client.emit('error', 'Lobby not found');
         }
     }
@@ -163,7 +133,6 @@ export class LobbyGateway implements OnGatewayConnection, OnGatewayDisconnect, O
         }
 
         lobby.isLocked = true;
-        this.logger.log(`Lobby ${accessCode} has been locked`);
         this.server.to(accessCode).emit('lobbyLocked', { accessCode, isLocked: true });
     }
 
@@ -181,7 +150,6 @@ export class LobbyGateway implements OnGatewayConnection, OnGatewayDisconnect, O
         }
 
         lobby.isLocked = false;
-        this.logger.log(`Lobby ${accessCode} has been unlocked`);
         this.server.to(accessCode).emit('lobbyUnlocked', { accessCode, isLocked: false });
     }
 
@@ -218,12 +186,12 @@ export class LobbyGateway implements OnGatewayConnection, OnGatewayDisconnect, O
         client.emit('avatarDeselected');
     }
 
-    afterInit() {
-        this.logger.log('LobbyGateway initialized.');
+    @SubscribeMessage('manualDisconnect')
+    handleManualDisconnect(@ConnectedSocket() client: Socket, @MessageBody() data: { isInGame: boolean } = { isInGame: false }) {
+        this.handlePlayerDisconnect(client, data.isInGame);
     }
 
     handleConnection(@ConnectedSocket() socket: Socket) {
-        this.logger.log(`User connected: ${socket.id}`);
         const accessCode = this.lobbyService.getLobbyIdByPlayer(socket.id);
 
         if (accessCode) {
@@ -233,12 +201,83 @@ export class LobbyGateway implements OnGatewayConnection, OnGatewayDisconnect, O
     }
 
     handleDisconnect(@ConnectedSocket() client: Socket) {
-        const accessCode = this.lobbyService.getLobbyIdByPlayer(client.id);
-        if (accessCode) {
-            this.lobbyService.leaveLobby(accessCode, client.id);
+        const accessCode = this.lobbyService.getRoomForPlayer(client.id);
+        const isInGame = this.gameSessionService.getGameSession(accessCode) !== undefined;
+        this.handlePlayerDisconnect(client, isInGame);
+    }
+
+    private handlePlayerDisconnect(client: Socket, isInGame: boolean = false) {
+        const accessCode = this.lobbyService.getRoomForPlayer(client.id);
+        const player = this.lobbyService.getPlayerBySocketId(client.id);
+        if (!player) {
+            this.handleFormDisconnect(client);
+            return;
+        }
+        if (isInGame) {
+            this.handleGamePlayerDisconnect(accessCode, player.name, client);
+        } else {
+            this.handleLobbyPlayerDisconnect(accessCode, player.name, client.id);
+        }
+        this.server.to(client.id).emit('updateUnavailableOptions', { avatars: [] });
+        client.leave(accessCode);
+        this.lobbyService.removePlayerSocket(client.id);
+    }
+
+    private handleGamePlayerDisconnect(accessCode: string, playerName: string, client: Socket) {
+        const playerAbandon = this.gameSessionService.getPlayers(accessCode).find((p) => p.name === playerName);
+        this.gameSessionService.handlePlayerItemReset(accessCode, playerAbandon);
+        this.gameCombatService.handleCombatSessionAbandon(accessCode, playerName);
+        this.gameSessionService.handlePlayerAbandoned(accessCode, playerName);
+        this.lobbyService.leaveLobby(accessCode, playerName, true);
+        const lobby = this.lobbyService.getLobby(accessCode);
+        const areAllVirtual = lobby.players.every((player) => player.isVirtual);
+        if ((lobby && lobby.players.length <= 1) || areAllVirtual) {
+            this.lobbyService.clearLobby(accessCode);
+            this.gameSessionService.deleteGameSession(accessCode);
+            this.accessCodesService.removeAccessCode(accessCode);
+            this.server.to(accessCode).emit('gameDeleted');
+            this.server.to(accessCode).emit('updateUnavailableOptions', { avatars: [] });
+        }
+        this.server.to(client.id).emit('updateUnavailableOptions', { avatars: [] });
+        this.server.to(accessCode).emit('game-abandoned', { player: playerAbandon });
+    }
+
+    private handleLobbyPlayerDisconnect(accessCode: string, playerName: string, clientId: string) {
+        const isAdminLeaving = this.lobbyService.isAdminLeaving(accessCode, playerName);
+        if (isAdminLeaving) {
+            this.server.to(accessCode).emit('adminLeft', {
+                playerSocketId: clientId,
+                message: "L'admin a quitté la partie, le lobby est fermé.",
+            });
+        }
+        const isLobbyDeleted = this.lobbyService.leaveLobby(accessCode, playerName);
+        const lobby = this.lobbyService.getLobby(accessCode);
+        if (lobby) {
+            lobby.waitingPlayers = lobby.waitingPlayers.filter((wp) => wp.socketId !== clientId);
+        }
+        if (isLobbyDeleted) {
+            this.server.to(accessCode).emit('lobbyDeleted');
+            this.server.to(accessCode).emit('updateUnavailableOptions', { avatars: [] });
+        } else if (lobby) {
+            const unavailableAvatars = [...lobby.players.map((p) => p.avatar), ...lobby.waitingPlayers.map((wp) => wp.avatar)];
+            this.server.to(accessCode).emit('updateUnavailableOptions', { avatars: unavailableAvatars });
             this.server.to(accessCode).emit('updatePlayers', this.lobbyService.getLobbyPlayers(accessCode));
         }
+    }
+
+    private async handleFormDisconnect(client: Socket) {
+        const accessCode = this.lobbyService.getRoomForPlayer(client.id);
+        if (!accessCode) return;
+        const lobby = this.lobbyService.getLobby(accessCode);
+        if (!lobby) return;
+        const waitingPlayerIndex = lobby.waitingPlayers.findIndex((wp) => wp.socketId === client.id);
+        if (waitingPlayerIndex === -1) return;
+        lobby.waitingPlayers.splice(waitingPlayerIndex, 1);
+        const unavailableAvatars = [...lobby.players.map((p) => p.avatar), ...lobby.waitingPlayers.map((wp) => wp.avatar)];
+        this.server.to(accessCode).emit('updateUnavailableOptions', {
+            avatars: unavailableAvatars,
+        });
+        client.leave(accessCode);
         this.lobbyService.removePlayerSocket(client.id);
-        this.logger.log(`User disconnected: ${client.id}`);
     }
 }
